fix websocket-based updates
On more recent versions of the websocket library `clients` cannot be
iterated over using a `for in` loop, the body will never run. Using
the `forEach` method fixes this issue.

Before this patch, the task and user lisk was not updated automatically (it
required a reload to refresh) because the websocket updates were not
broadcasted correctly to all connected clients.